backend/requirements.txt
# Simple CrewAI Education Platform Requirements

# Core Framework
fastapi==0.115.14
uvicorn[standard]==0.35.0

# Database
SQLAlchemy
psycopg2-binary==2.9.10
alembic==1.16.2

# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20

# CrewAI and AI dependencies

openai==1.93.0
anthropic==0.55.0

# Utility dependencies
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1
aiohttp==3.10.11
aiofiles==24.1.0

# PDF processing (for scenario uploads)
pdfplumber==0.11.7
pdfminer.six==20250506

# Development dependencies (optional)
pytest==8.4.1

# For deployment
PyYAML==6.0.2